+++ operator/xover/bitstring_mater.go
@@ -15,7 +15,7 @@ type BitstringMater struct{}
 //
 // parent1 and parent2 are the two individuals that provides the source
 // material for generating offspring.
-func (BitstringMater) Mate(p1, p2 *bitstring.Bitstring, nxpts int64, rng *rand.Rand) []*bitstring.Bitstring {
+func (BitstringMater) Mate(p1, p2 *bitstring.Bitstring, nxpts int, rng *rand.Rand) []*bitstring.Bitstring {
 	if p1.Len() != p2.Len() {
 		panic("Cannot mate parents of different lengths")
 	}
@@ -23,7 +23,7 @@ func (BitstringMater) Mate(p1, p2 *bitstring.Bitstring, nxpts int64, rng *rand.R
 	off2 := bitstring.Clone(p2)
 
 	// Apply as many crossovers as required.
-	for i := int64(0); i < nxpts; i++ {
+	for i := 0; i < nxpts; i++ {
 		// Cross-over index is always greater than zero and less than the
 		// length of the parent so that we always pick a point that will
 		// result in a meaningful crossover.
